+++ Cargo.toml
@@ -8,7 +8,6 @@ edition = "2018"
 modular-bitfield = "0.11"
 heapless = "0.6"
 crc-any = { version = "2.3", default-features = false }
-proptest = "1.0"
 
 [dev-dependencies]
 socketcan = "1.7"
