+++ src/session_id/session_kind.rs
@@ -90,7 +90,7 @@ pub fn can_id_for_session_kind(kind: SessionKind, priority: TransferPriority) ->
     }
 }
 
-#[cfg(any(debug_assertions, test))]
+#[cfg(test)]
 pub mod strategy {
     use super::super::node_id::strategy::node_id;
     use super::super::service_id::strategy::service_id;
