--- tests/common/mod.rs
@@ -1,2 +1,0 @@
-pub mod can_frame;
-pub mod can_tx;
